Reverting java snippet and example back to original
