--- snippets-examples/robot-3270-snippet/.gitignore
@@ -1,1 +1,0 @@
-/target/
