--- snippets-examples/robot-3270-snippet/src/main/resources/META-INF/IBM3270ManagerRobot.json
@@ -1,1 +1,0 @@
-{"id":"IBM3270ManagerRobot","robotType":"ROBOT","mainClass":"com.appian.rpa.snippets.examples.IBM3270ManagerRobot","actions":["init","open3270","login","goToNetView","changePassword","close3270","end"],"conditionalActions":["startUp"],"fieldLinks":[],"actionDetails":[{"id":"init","implementation":"/**\n\t * Action 'Init'.\n\t * <p>\n\t * Initializes Jidoka modules. Instances of the emulator type passed as a\n\t * parameter are loaded\n\t */\npublic void init() {\n    try {\n        server.debug(\"Robot initialized\");\n    } catch (Exception e) {\n        throw new  JidokaFatalException(\"Error initializing\");\n    }\n}"},{"id":"open3270","implementation":"/**\n\t * Action 'Open 3270' terminal\n\t */\npublic void open3270() {\n    server.info(\"Opening 3270 terminal\");\n    intanceIBM3270AppManager = IBM3270AppManager.openIBM3270(\"C:\\\\Program Files (x86)\\\\IBM\\\\Personal Communications\\\\private\\\\GSO1tls-MAI.ws\", \".*Session.*\", this, \"PCOMM\");\n    client.pause(1000);\n    server.sendScreen(\"Screenshot after opening the terminal\");\n}"},{"id":"login","implementation":"/**\n\t * Go to the page indicated as parameter\n\t * \n\t * @param page Name of the page in the menu\n\t * @throws JidokaException\n\t */\npublic void login() throws JidokaException {\n    intanceIBM3270AppManager.write(\"==>\", 4, 0, \"M\", 3);\n    intanceIBM3270AppManager.control();\n    client.pause(1000);\n    server.sendScreen(\"Screenshot Access Menu\");\n    intanceIBM3270AppManager.write(\"==>\", 4, 0, \"3\", 3);\n    intanceIBM3270AppManager.control();\n    server.sendScreen(\"Screenshot Login\");\n    client.pause(10000);\n    server.debug(\"Existe texto? \" + intanceIBM3270AppManager.existText(\"TSO DE SAQS\"));\n}"},{"id":"goToNetView","implementation":"/**\n\t * Action 'Go to NetView' page.\n\t * \n\t * @throws JidokaException\n\t */\npublic void goToNetView() throws JidokaException {\n    server.sendScreen(\"Screenshot before moving to NetView page\");\n    //currentPage = ((WelcomePage) currentPage).goToPage(\"NETVIEW\");\n    server.sendScreen(\"Moved to NetView page\");\n}"},{"id":"changePassword","implementation":"/**\n\t * Action 'Change password'.\n\t * \n\t * @throws JidokaException\n\t */\npublic void changePassword() throws JidokaException {\n    server.info(\"Change NetView Password\");\n}"},{"id":"close3270","implementation":"/**\n\t * Action 'Close 3270'.\n\t * @throws IOException \n\t */\npublic void close3270() throws IOException {\n    server.info(\"Closing IBM3270 terminal\");\n    intanceIBM3270AppManager.close();\n}"},{"id":"end","implementation":"/**\n\t * Action 'End'.\n\t */\npublic void end() {\n}"}],"conditionalActionDetails":[{"id":"startUp","implementation":"/**\n\t * Startup method. This <code>startUp</code> method is called prior to calling\n\t * the first workflow method defined\n\t */\n@SuppressWarnings(\"unchecked\")\n@Override\npublic boolean startUp() throws Exception {\n    server = (IJidokaServer<Serializable>) JidokaFactory.getServer();\n    client = IClient.getInstance(this);\n    return true;\n}"}],"sdkMethods":[]}
