+++ snippets-examples/robot-snippet-3270/src/main/java/com/appian/rpa/snippets/examples/pages/ConstantsTexts.java
@@ -1,0 +1,45 @@
+package com.appian.rpa.snippets.examples.pages;
+
+
+/**
+ * Class with unique texts on each page
+ */
+public class ConstantsTexts {
+
+	/**
+	 * Univocal text for Welcome Page
+	 */
+	public static final String WELCOME_UNIVOCAL_TEXT = "Welcome";
+	
+	/**
+	 * Univocal text for Netview Pages
+	 */
+	public static final String NETVIEW_UNIVOCAL_TEXT = "V5";
+	
+	/**
+	 * Univocal text for Change Password File
+	 */
+	public static final String PWD_UNIVOCAL_TEXT = "VERIFY";
+	
+	/**
+	 * Univocal text for Change Password File
+	 */
+	public static final String INVALID_USER_UNIVOCAL_TEXT = "INVALID OPERATOR";
+	
+	/**
+	 * Operator Id
+	 */
+	public static final String TEST_ID_OPERATOR = "OPER1";
+	
+	/**
+	 * Operator Password
+	 */
+	public static final String TEST_PWD_OPERATOR = "ZZZ";
+	
+	/**
+	 * Operator New Password
+	 */
+	public static final String TEST_NEW_PWD_OPERATOR = "YYY";
+		
+
+}
