+++ snippets-libraries/snippet-3270-manager/pom.xml
@@ -7,54 +7,50 @@
 	<artifactId>snippet-3270-manager</artifactId>
 	<version>7.5.0</version>
 	<packaging>jar</packaging>
+
 	<properties>
 		<maven.compiler.target>1.8</maven.compiler.target>
 		<maven.compiler.source>1.8</maven.compiler.source>
 	</properties>
-  <parent>
-    <groupId>com.novayre.jidoka.robot</groupId>
-    <artifactId>jidoka-robot-parent</artifactId>
-    <version>7.6.0</version>
-  </parent>
-  <profiles>
-    <profile>
-      <id>jidoka-repo</id>
-      <activation>
-        <activeByDefault>true</activeByDefault>
-      </activation>
-      <distributionManagement>
-        <repository>
-          <id>jidoka</id>
-          <name>jidoka</name>
-          <url>https://lincolndev.appiancloud.com/rpa/repo/</url>
-        </repository>
-      </distributionManagement>
-      <repositories>
-        <repository>
-          <releases>
-            <enabled>true</enabled>
-          </releases>
-          <snapshots>
-            <enabled>true</enabled>
-          </snapshots>
-          <id>jidoka</id>
-          <url>https://lincolndev.appiancloud.com/rpa/repo/</url>
-        </repository>
-      </repositories>
-      <pluginRepositories>
-        <pluginRepository>
-          <releases>
-            <enabled>true</enabled>
-          </releases>
-          <snapshots>
-            <enabled>true</enabled>
-          </snapshots>
-          <id>jidoka</id>
-          <url>https://lincolndev.appiancloud.com/rpa/repo/</url>
-        </pluginRepository>
-      </pluginRepositories>
-    </profile>
-  </profiles>
+
+	<parent>
+		<groupId>com.novayre.jidoka.robot</groupId>
+		<artifactId>jidoka-robot-parent</artifactId>
+		<version>7.5.0</version>
+	</parent>
+
+	<profiles>
+			<profile>
+			<id>rpa-coe-lab</id>
+			<activation>
+				<activeByDefault>true</activeByDefault>
+			</activation>
+			<distributionManagement>
+				<repository>
+					<id>rpa-coe-lab</id>
+					<url>https://rpa-coe-lab.appiancloud.com/rpa/repo/</url>
+				</repository>
+			</distributionManagement>
+			<repositories>
+				<repository>
+					<releases />
+					<snapshots />
+					<id>rpa-coe-lab</id>
+					<url>https://rpa-coe-lab.appiancloud.com/rpa/repo/</url>
+				</repository>
+			</repositories>
+			<pluginRepositories>
+				<pluginRepository>
+					<releases />
+					<snapshots />
+					<id>rpa-coe-lab</id>
+					<url>https://rpa-coe-lab.appiancloud.com/rpa/repo/</url>
+				</pluginRepository>
+			</pluginRepositories>
+		</profile>
+	</profiles>
+
+
 	<build>
 		<plugins>
 			<!-- we want JDK 1.8 source and binary compatiblility -->
@@ -89,4 +85,5 @@
 			<version>1.3</version>
 		</dependency>
 	</dependencies>
+
 </project>
